requirements.txt
scikit-learn
joblib
pandas
flask_sqlalchemy 
flask-mail